Reducing data sizes

+++ horn.cpp
@@ -3,7 +3,7 @@
 
 Horn::Horn() : horn(OFF) {}
 
-void Horn::setup(Ui::MainWindow *ui, int height, int width) {
+void Horn::setup(Ui::MainWindow *ui, uint8_t height, uint8_t width) {
   this->ui = ui;
 
   horn_icon = QPixmap(":/icons/horn.png");
